Move plugins menu tests from '_qt.menus._tests' to '_qt._qapp_model._tests'
